+++ Sources/libcsuper/csu_struct.h
@@ -98,5 +98,6 @@ void addDistributorCsuStruct(csuStruct *ptr_csu_struct, char *distributor_name);
 int exceedMaxNumber(csuStruct *ptr_csu_struct);
 int maxNbTurn(csuStruct *ptr_csu_struct);
 int searchPlayerIndex(csuStruct *ptr_csu_struct, char *player_name);
+int differentsPlayerName(csuStruct *ptr_csu_struct);
 
 #endif