+++ Changelog.md
@@ -2,8 +2,10 @@ Csuper
 ======
 
 4.0.0
-* Change the dispplay of a game configuration with yes or no instead of 1 and 0
-* Correction of a bug which display the distrubutor when there is no distributor
+
+* Change the display of a game configuration with yes or no instead of 1 and 0
+* Correction of a bug which display the distributor when there is no distributor
+* A few optimization
 
 2.4.0
 
@@ -14,8 +16,13 @@ Libcsuper
 =========
 
 4.0.0
+
 * Add a function to transform integer to yes or no
 * Add a function which calculate the index of the array from the position in the game
+* The function changeSystemPath check if the new path is valid
+* New function checkPath
+* New function getFolderFromFilename
+* New function differentsPlayerName
 
 2.4.0